Adds missing cast.

+++ lib/DxcSupport/HLSLOptions.cpp
@@ -555,7 +555,8 @@ int ReadDxcOpts(const OptTable *optionTable, unsigned flagsToInclude,
         errors << "/enable-templates is not supported with HLSL Version " << opts.HLSLVersion;
 
       if (opts.EnableUnions)
-        errors << "/enable-unions is not supported with HLSL Version " << opts.HLSLVersion;
+        errors << "/enable-unions is not supported with HLSL Version "
+               << (unsigned long)opts.HLSLVersion;
 
       if (opts.StrictUDTCasting)
         errors << "/enable-udt-casting is not supported with HLSL Version " << opts.HLSLVersion;
